fix user auth

+++ backend/models.py
@@ -1,5 +1,6 @@
 # backend/models.py
 from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, Float, String  
 from sqlalchemy.orm import relationship
 from .database import Base
 
@@ -11,13 +12,23 @@ class User(Base):
     hashed_password = Column(String)
     orders = relationship("Order", back_populates="owner")
 
+# class Product(Base):
+#     __tablename__ = "products"
+
+#     id = Column(Integer, primary_key=True, index=True)
+#     name = Column(String, index=True)
+#     description = Column(String)
+#     price = Column(Integer)
+
+
 class Product(Base):
     __tablename__ = "products"
 
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, index=True)
     description = Column(String)
-    price = Column(Integer)
+    price = Column(Float)  # Altere de Integer para Float
+
 
 class Order(Base):
     __tablename__ = "orders"
